requirements.txt
Flask_Mail==0.9.1
Flask_Bcrypt==0.7.1
flask_mongoengine==0.9.5
Flask_RESTful==0.3.8
mongoengine==0.19.1
Flask_JWT_Extended==3.24.1
Flask==1.1.1
flask_security==3.0.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pytest==5.4.1
